% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_scores.R
\name{assign_pval}
\alias{assign_pval}
\title{assign_pval}
\usage{
assign_pval(MIC_scores, nb_cells)
}
\arguments{
\item{MIC_scores}{numeric vector of MIC scores}

\item{nb_cells}{nb of cells in the gene expression matrix}
}
\value{
Returns a vector of p-values
}
\description{
Assigns p-values to MIC scores
}
\details{
Since the p-value of a given MIC score cannot be computed
by a simple mathemical function, \code{assigns_pval} reads
the p-value table generated by the authors of the MINE
statistics (see \url{http://www.exploredata.net/Downloads/P-Value-Tables})
at the closest \emph{sample_size} to \emph{nb_cells}.

For each \emph{MIC score}, \code{assign_pval} will then
retrieve the corresponding p-value from this table.
}
