% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clara_clustering}
\alias{clara_clustering}
\title{clara_clustering}
\usage{
clara_clustering(scores_tab, k)
}
\arguments{
\item{scores_tab}{data.table; the table of scores}

\item{k}{desired number of clusters}
}
\value{
a vector corresponding to the partition of the table of scores
}
\description{
uses clara algorithm to partition the table of scores into k clusters
around medoids (a more robust version of K-means)
}
