% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMatrix.R
\name{collapseMatrix}
\alias{collapseMatrix}
\title{Collapse matrix model into propagule, pre-reproductive, reproductive and
post-reproductive stages.}
\usage{
collapseMatrix(matU, matF, collapse)
}
\arguments{
\item{matU}{survival matrix}

\item{matF}{fecundity matrix}

\item{collapse}{(character) vector indicating which stages must be collapsed
and retained to give matrix containing combined propagule, pre-reproductive,
reproductive and post-reproductive stages.}
}
\value{
(list) collapsed matrices \code{matA}, \code{matU}, \code{matF} and
character vector \code{matrixStages} specifying which matrix stages are
included in the collapse model.
}
\description{
Note: only valid for models without clonality.
}
\examples{
matU <- matrix(c(0.2581, 0.1613, 0.1935, 0.2258, 0.1613, 0.0408, 0.2857,
                 0.4286, 0.102, 0.0816, 0.0385, 0.0385, 0.2692, 0.2308,
                 0.3462, 0, 0.0625, 0.125, 0.25, 0.5625, 0.1061, 0.1608,
                 0.2637, 0.1801, 0.2058),
                 nrow = 5, byrow = FALSE)
matF <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 2.75, 1.75, 0, 0),
                 nrow = 5, byrow = FALSE)
collapse1 <- c("1-2","3-4","5")
(out <- collapseMatrix(matU, matF, collapse1))
eigen(out$matA)
eigen(out$matU)
eigen(out$matF)

#collapse2 <- c("1-2","3-4-5")
#collapse3 <- c("1-2-3-4-5")
}
