% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{compute_density}
\alias{compute_density}
\title{compute_density}
\usage{
compute_density(x)
}
\arguments{
\item{x}{matrix: the matrix of scores (without the gene names)}
}
\value{
a numeric vector of densities
}
\description{
Uses Multidimensional Kernel Density Estimation to
estimate the density of each association in the space of the scores
}
