% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_and_manip_scores.R
\name{compute_scores}
\alias{compute_scores}
\title{compute_scores}
\usage{
compute_scores(x, n.cores = 1)
}
\arguments{
\item{x}{a gene expression matrix (row.names corresponding to cells, col.names to genes)}

\item{n.cores}{nb of cores to be used for parallel processing}
}
\description{
Computes the table of association and characterization scores
}
\details{
Structure of the returned table of scores
\itemize{
  \item{column 1:2} {- the two genes whose assocation is being tested}
  \item{column 3} {- MIC score, which is the strength of the association between the 2 genes}
  \item{column 4:n} {- characterization scores, informing about the "shape" of the association}
}
}
