% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_exploration.R
\name{explore_clusters}
\alias{explore_clusters}
\title{explore_clusters}
\usage{
explore_clusters(scores_tab, partition, gene_exp_mat,
  sampling_method = "downsampling", target_ratio = 0.3, target_nb = NULL,
  output_directory = ".")
}
\arguments{
\item{scores_tab}{data.table; the table of scores}

\item{partition}{categorial vector; the partition of the \code{scores_tab} into k clusters.
\code{length(partition)} must be equal to \code{nrow(scores_tab)}}

\item{gene_exp_mat}{matrix: the gene expression matrix used for computing the scores
(rownames corresponding to genes and colnames to cells)}

\item{sampling_method}{method to be used for selecting a few associations in each branch.

Methods available:
\itemize{
  \item{\code{"downsampling"}}
  \item{\code{"lowsampling"}}
  \item{\code{"upsampling"}}
  \item{\code{"centroids"}}
  \item{\code{"medoids"}}
 }

Method recommanded: \code{"downsampling"}}

\item{target_ratio}{number; fraction of associations to be sampled in each tree's node
. Argument \code{target_ratio} will be ignored if argument \code{target_nb} is provided}

\item{target_nb}{number of associations to be sampled in each tree's node}

\item{output_directory}{where to create the arborescence of directories for each cluster}
}
\description{
Given a table of scores and its partition into k clusters,
the function \code{explore_clusters} proposes to explore the types
of associations lying within each cluster.

To do so, the function measures the dispersion of the associations of each
cluster in the space of the scores and depicts this heterogeneity
as a tree structure (using ElPiGraph.R package).
For instance, nodes along a same branch will be expected to exhibit
similar types of associations.

Then, the function will visit the tree of each cluster to create an arborescence of
directories starting from the root (the root is defined
as the branch which is the closest to the centroid of the table of scores)
and expanding from adjacent branches to adjacent branches.

Inside each directory corresponding to a branch (or a branching point),
the function will produce the scatterplots of a few sampled assocations in order
for the user to get a glimpse of what type of associations each branch is related to.

By visiting this arborescence of directories and visualizing
some of the scatterplots of the associations each branch is related to,
it enables one to somehow "wander" along a cluster's tree and appreciate the
"evolution" of the "shape/type" of the associations from branches to branches
}
