% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVitalRates.R
\name{extractVitalRates}
\alias{extractVitalRates}
\title{Function to extract the vital rates of a 4x4 matrix that has been
collapsed into propagule, pre-, reprod, and post-reprod}
\usage{
extractVitalRates(matU, matF, collapse)
}
\arguments{
\item{matU}{a matrix without reproduction and clonality}

\item{matF}{a matrix, just reproduction}

\item{collapse}{vector of stages}
}
\description{
FIXME: change to allow non 4 by 4 matrices
}
\examples{
matU <- matrix(c(0.2581, 0.1613, 0.1935, 0.2258, 0.1613, 0.0408, 0.2857,
                 0.4286, 0.102, 0.0816, 0.0385, 0.0385, 0.2692, 0.2308,
                 0.3462, 0, 0.0625, 0.125, 0.25, 0.5625, 0.1061, 0.1608,
                 0.2637, 0.1801, 0.2058),
                 nrow = 5, byrow = FALSE)
matF <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 2.75, 1.75, 0, 0),
                 nrow = 5, byrow = FALSE)
collapse <- c("1-2", "3", "4", "5")
xx <- collapseMatrix(matU, matF, collapse)
extractVitalRates(xx$matU, xx$matF, collapse)
}

