% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_scores_table.R
\name{filter_scores}
\alias{filter_scores}
\title{filter_scores}
\usage{
filter_scores(x, on = "MIC", pval = NULL, thresh = 0.4)
}
\arguments{
\item{x}{data.frame: the table of scores}

\item{on}{variable used for the threshold (accepts only \code{"MIC"} or \code{"pval"})}

\item{pval}{numeric vector of p-values corresponding to the MIC scores in \code{x} (only if \code{on = "pval"})}

\item{thresh}{value of the threshold}
}
\value{
Returns a filtered table of scores
}
\description{
Filters a table of scores based on a threshold of significance defined either by a MIC score, or by a p-value
}
\details{
if \code{on = "MIC"}, \code{filter_scores} will retain any association whose MIC >= \code{threshold}

else if \code{on = "pval"}, \code{filter_scores} will retain any association whose p-value <= \code{threshold}
}
\examples{
\dontrun{
scores_tab <- compute_scores(my_gene_exp_matrix, n.cores = 6)
pvalues <- assign_pval(scores_tab$`MIC (strength)`, nb_cells = 96)
signif_scores_tab <- filter_scores(scores_tab, on = "pval", 
                                   pval = pvalues, threshold = 0.05)
}                                    

}
