% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{get_classified_asso}
\alias{get_classified_asso}
\title{get_classified_asso}
\usage{
get_classified_asso(trees_info, classification_file, scores_tab)
}
\arguments{
\item{trees_info}{the trees info returned by \code{explore_clusters} function}

\item{classification_file}{the filled in classification template produced by \code{explore_clusters}
\code{length(partition)} must be equal to \code{nrow(scores_tab)}}

\item{scores_tab}{the table of scores}
}
\value{
A list of tables, each of them corresponding to the associations of a specific classification
}
\description{
After having called the \code{explore_clusters}, this function proposes
to retrieve ALL the associations corresponding to any assigned classification in the 
'associations_classification_template.csv' file.
}
