% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_optimal_k}
\alias{get_optimal_k}
\title{get_optimal_k}
\usage{
get_optimal_k(x, ix = NULL, cutoff = 0.8, nrep = 20)
}
\arguments{
\item{x}{data.table; the table of scores}

\item{ix}{boolean vector of indexes to which x must be reduced (sampling output)}

\item{cutoff}{cutoff for the prediction strength (recommanded to range from 0.8 up to 0.9)}

\item{nrep}{nb of times the clustering is performed for each value of k (bootstrapping)}
}
\value{
the optimal number of clusters
}
\description{
a wrapper to \code{fpc::prediction.strength} function. This function is based on
consensus clustering using clara algorithm for different k (2 to 10 precisely).

The prediction strength is defined according to Tibshirani and Walther (2005) who recommand to
choose as optimal number of clusters the largest number of clusters that leads to
a prediction strength above a \code{cutoff} of 0.8 or 0.9.
}
