\name{Mage-package}
\alias{Mage-package}
\alias{Mage}
\docType{package}
\title{
\packageTitle{Mage}
}
\description{
\packageDescription{Mage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Mage}
\packageIndices{Mage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Mage}

Maintainer: \packageMaintainer{Mage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
