% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{phenograph_clustering}
\alias{phenograph_clustering}
\title{phenograph_clustering}
\usage{
phenograph_clustering(scores_tab, ngb = 50)
}
\arguments{
\item{scores_tab}{data.table; the table of scores}

\item{ngb}{number of nearest neighbours}
}
\value{
a vector corresponding to the partition of the table of scores
}
\description{
uses the phenograph algorithm to perform an automated clustering
of the table of scores
}
