% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_scores_table.R
\name{plot_scores_relationships}
\alias{plot_scores_relationships}
\title{plot_scores_relationships}
\usage{
plot_scores_relationships(x, pdf_filename = NULL)
}
\arguments{
\item{x}{data.frame: the table of scores}

\item{pdf_filename}{path to the pdf file (optional)}
}
\description{
Produce the scatterplots of each pair of scores and display them
as a symetric matrix of plots. 

This function is meant to visualize
the kind of relationships that exist between the different scores.
}
