% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeMatrix.R
\name{rearrangeMatrix}
\alias{rearrangeMatrix}
\title{Re-arrange matrix stages so that all inter-reproductive and non-reproductive
stages fall in the final rows/columns of the matrix. This is a preparatory
step to collapsing the matrix model into combined prop, pre-rep, rep and
post-rep stages.}
\usage{
rearrangeMatrix(matU, matF, matFmu)
}
\arguments{
\item{matU}{survival matrix}

\item{matF}{fecundity matrix}

\item{matFmu}{mean fecundity matrix}
}
\description{
Re-arrange matrix stages so that all inter-reproductive and non-reproductive
stages fall in the final rows/columns of the matrix. This is a preparatory
step to collapsing the matrix model into combined prop, pre-rep, rep and
post-rep stages.
}
\examples{
## FIXME: find an example containing non-reproductive stages
matU <- matrix(c(0.2581, 0.1613, 0.1935, 0.2258, 0.1613, 0.0408, 0.2857,
                 0.4286, 0.102, 0.0816, 0.0385, 0.0385, 0.2692, 0.2308,
                 0.3462, 0, 0.0625, 0.125, 0.25, 0.5625, 0.1061, 0.1608,
                 0.2637, 0.1801, 0.2058),
                 nrow = 5, byrow = FALSE)
matF <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 2.75, 1.75, 0, 0),
                 nrow = 5, byrow = FALSE)
matFmu <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                   0, 0, 0, 0, 1.0409262, 0.5040727, 0.016433, 0.06956801),
                 nrow = 5, byrow = FALSE)
rearrangeMatrix(matU, matF, matFmu)
}
