% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample}
\alias{sample}
\title{sample}
\usage{
sample(x, partition = NULL, methods = "downsampling", target_ratio = 0.1,
  target_nb = NULL)
}
\arguments{
\item{x}{data.frame, a table of scores}

\item{partition}{vector containing the partition of table \code{x}.

Each element of the partition vector must be a cluster ID and length(partition) must be equal to nrow(x).

If \code{partition} is passed to this function, each cluster will be sampled instead of the whole table.}

\item{methods}{method or vector of methods to be used for the sampling.

Methods available:
\itemize{
  \item{\code{"downsampling"}}
  \item{\code{"lowsampling"}}
  \item{\code{"upsampling"}}
  \item{\code{"centroids"}}
  \item{\code{"medoids"}}
 }

See Details to have the description of the methods}

\item{target_ratio}{percentage of gene pairs to be retained in \code{x}
(if partition is NULL) or in each cluster of \code{x} (if partition is NOT NULL)

argument \code{target_ratio} will be ignored if argument \code{target_nb} is provided}

\item{target_nb}{number of gene pairs to be retained in \code{x} (if partition is NULL) or in each cluster of \code{x} (if partition is NOT NULL)}
}
\description{
Sample the table of scores or clusters of it
according to different methods of selection
}
