% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show_associations}
\alias{show_associations}
\title{show_associations}
\usage{
show_associations(x, y, partition, partition_ixs, output_dir)
}
\arguments{
\item{x}{data.frame; the table of scores}

\item{y}{matrix; the gene expression matrix
(rownames corresponding to genes and colnames to cells)}

\item{partition}{vector; partition of \code{x} into k clusters}

\item{partition_ixs}{list of n-by-m matrices of boolean indexes
resulting from \code{sample} function
where \code{n} is the number of individuals in a given cluster of the \code{partition}
and \code{m} is is the number of methods used for the sampling}

\item{output_dir}{path to a directory where to store all the plots}
}
\description{
Produces the scatterplot of each sampled association between 2 genes inside
each cluster of a partition
}
