% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show_heatmap}
\alias{show_heatmap}
\title{show_heatmap}
\usage{
show_heatmap(x, partition, ix = NULL, pdf_filename = NULL)
}
\arguments{
\item{x}{data.frame; the standardize table of scores (consider using \code{standardize_scores} function)}

\item{partition}{vector containing the partition of the table \code{x}.}

\item{ix}{boolean vector of indexes to which the heatmap must be reduced (\code{sample} output).
A hundred representative individuals by cluster (downsampling) should be enough to get a glimpse
of what the cluster looks like in term of profile of scores ...

\code{ix} is a recommanded argument for a large table of score (greater than 1000 individuals)}

\item{pdf_filename}{path to a pdf file which must store the plot (optional)}
}
\description{
Show the heatmap of scores for each cluster in the partition
}
