% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show_reduced_space}
\alias{show_reduced_space}
\title{show_reduced_space}
\usage{
show_reduced_space(x, x_ixs = NULL, partition = NULL,
  partition_ixs = NULL, PCs = 1:4, output_dir = NULL)
}
\arguments{
\item{x}{data.frame; the table of scores}

\item{x_ixs}{(optional) list containing a unique n-by-m matrix of boolean indexes
resulting from the \code{sample} function
where \code{n} is the number of individuals in \code{x}
and \code{m} is the number of methods used for the sampling}

\item{partition}{(optional) vector; partition of \code{x} into k clusters}

\item{partition_ixs}{(optional) list of n-by-m matrices of boolean indexes
resulting from \code{sample} function
where \code{n} is the number of individuals in a given cluster of the \code{partition}
and \code{m} is is the number of methods used for the sampling}

\item{PCs}{numeric vector of Principal Compenents to be considered (max = 5)}

\item{output_dir}{(optional) path to a directory where to store all the plots}
}
\value{
Returns a list of plots
}
\description{
Reduce the space of the scores to its first principal components
in order to display on it the associated gene pairs.
}
\details{
the function \code{show_reduced_space} returns a graphical object
composed of 2 up to 5 elements:
\describe{
  \item{PCA_information}{The graphical informations about the PCA}
  \item{population}{Show all the gene pairs in the reduced space}
  \item{population_samples}{Show the sampled gene pairs within the population
  (\code{ixs} must be provided)}
  \item{clusters}{Show the different clusters of the population
  (\code{partition} must be provided)}
  \item{clusters_samples}{Show the different sampled gene pairs within each clusters
  (\code{partition_ixs} must be provided)}
}
}
