% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_scores.R
\name{adjust_pval}
\alias{adjust_pval}
\title{adjust_pval}
\usage{
adjust_pval(pval)
}
\arguments{
\item{pval}{numeric vector of p_values}
}
\value{
Returns a vector of adjusted p-values
}
\description{
Adjust p-values using FDR method (Benjamini & Hochberg, 1995)
}
