% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores_and_pval.R
\name{compute_scores}
\alias{compute_scores}
\title{compute_scores}
\usage{
compute_scores(x, n.cores = 1)
}
\arguments{
\item{x}{matrix: a gene expression matrix
(rownames corresponding to genes and colnames to cells)}

\item{n.cores}{integer: nb of cores to be used for parallel processing}
}
\value{
Returns a table with the following columns:
\describe{
  \item{1:2}{Names of the two genes whose assocation is being assessed}
  \item{3}{MIC score, \emph{i.e} the strength of the association between the 2 genes}
  \item{4:n}{Characterization scores, informing about the "shape" of the association}
}
}
\description{
Computes the table of association and characterization scores
}
