% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_scores.R
\name{filter_scores}
\alias{filter_scores}
\title{filter_scores}
\usage{
filter_scores(x, on = "MIC", pval = NULL, threshold = 0.4)
}
\arguments{
\item{x}{data.frame, table of scores}

\item{on}{variable name used for the threshold (set to either \code{"MIC"} or \code{"pval"})}

\item{pval}{a numeric vector of pval (only if \code{on = "pval"})}

\item{threshold}{the value of the threshold}
}
\value{
Returns a filtered table of scores
}
\description{
Filter table of scores based on a threshold of significance defined either by a MIC score, or by a p-value
}
