\name{Mage-package}
\alias{Mage-package}
\alias{Mage}
\docType{package}
\title{
\packageTitle{Mage}
}
\description{
\packageDescription{Mage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Mage}
\packageIndices{Mage}
This package provides tools for interacting with the COMPADRE and COMADRE Matrix Databases.
}

\author{
\packageAuthor{Mage}

Maintainer: \packageMaintainer{Mage}
}

\references{
Salguero-Gomez, R., Jones, O. R., Archer, C. R., Buckley, Y. M., Che-Castaldo, J., Caswell, H., Hodgson, D., Scheuerlein, A., Conde, D. A., Brinks, E., de Buhr, H., Farack, C., Gottschalk, F., Hartmann, A., Henning, A., Hoppe, G., Roemer, G., Runge, J., Ruoff, T., Wille, J., Zeh, S., Davison, R., Vieregg, D., Baudisch, A., Altwegg, R., Colchero, F., Dong, M., de Kroon, H., Lebreton, J.-D., Metcalf, C. J. E., Neel, M. M., Parker, I. M., Takada, T., Valverde, T., Velez-Espino, L. A., Wardle, G. M., Franco, M., Vaupel, J. W. (2015), The compadre Plant Matrix Database: an open online repository for plant demography. Journal of Ecology, 103: 202-218. doi: 10.1111/1365-2745.12334

Salguero-Gomez, R., Jones, O. R., Archer, C. R., Bein, C., de Buhr, H., Farack, C., Gottschalk, F., Hartmann, A., Henning, A., Hoppe, G., Roemer, G., Ruoff, T., Sommer, V., Wille, J., Voigt, J., Zeh, S., Vieregg, D., Buckley, Y. M., Che-Castaldo, J., Hodgson, D., Scheuerlein, A., Caswell, H., Vaupel, J. W. (2016), COMADRE: a global data base of animal demography. Journal of Animal Ecology, 85: 371-384. doi: 10.1111/1365-2656.12482

}

\keyword{ package }

\seealso{

}

\examples{

}
