% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodStages.R
\name{reprodStages}
\alias{reprodStages}
\title{Reproductive stages}
\usage{
reprodStages(matF, matFmu, post, maxRep, matrixStages = NULL)
}
\arguments{
\item{matF}{a matrix, just reproduction}

\item{matFmu}{a matrix}

\item{post}{post something}

\item{maxRep}{maximum xxx}

\item{matrixStages}{stages}
}
\description{
Determine pre-reproductive, reproductive and post-reproductive stages
from a matrix model which pre-formats them to use them as the "collapse"
argument of the function above to collapse matrices
}
\details{
FIXME: ROB says possibly put output as a named list instead of vector
}
\examples{
# reprodStages(x, y, z, a)
}

