% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodStages.R
\name{reprodStages}
\alias{reprodStages}
\title{Group the stages of a matrix model into propagule, pre-reproductive,
reproductive and post-reproductive stages.}
\usage{
reprodStages(matF, matFmu, post, maxRep, matrixStages, includeProp = TRUE,
  includePost = TRUE)
}
\arguments{
\item{matF}{(matrix) fecundity matrix, rearranged so that non-reproductive
stages are in the final rows/columns}

\item{matFmu}{(matrix) mean fecundity matrix, rearranged so that
non-reproductive stages are in the final rows/columns}

\item{post}{(integer) indicator for post-reproductive stages}

\item{maxRep}{(integer) maximum number of reproductive stages}

\item{matrixStages}{(character) vector of stages, values are "prop"
(propagule), "active", and "dorm" (dormant)}

\item{includeProp}{(logical) include propagule stage. default: \code{TRUE}.
if \code{TRUE}, propagule stage (if present) is given back in result. If
\code{FALSE}, it's included into the pre-reproductive stage}

\item{includePost}{(logical) include post-reproductive stage. default:
\code{TRUE}. if \code{TRUE}, post-reproductive stage (if present) is given
back in result. If \code{FALSE}, it's included into the reproductive
stage}
}
\description{
Assumes that fecundity and mean fecundity matrices have been rearranged so
that non-reproductive stages are in the final rows/columns.
}
\details{
Output indicates groupings to be used when collapsing the matrix model.

Note: dormant stages not handled.

FIXME: ROB says possibly put output as a named list instead of vector
FIXME: can we combine propagule and pre-reproductive?
FIXME: what about dormant stages?
FIXME: once above solved, then fix logic internally

FIXME: Scott and Tamora added the extra parameter flags for switching whether
to include propagule and post-reproductive stages but these are not used as
yet. On reflection this is probably going too far and we should let the
matrix structure guide the process. However, in animals we may wish to
combine rep and post-rep to obtain a two-stage matrix model.
}
\examples{
matF <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 2.75, 1.75, 0, 0),
                 nrow = 5, byrow = FALSE)
matFmu <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                   0, 0, 0, 0, 0, 1.0409262, 0.5040727, 0.016433,
                   0.06956801),
                   nrow = 5, byrow = FALSE)
post <- numeric(0)
maxRep <- 5
matrixStages <- c("dorm", "active", "active", "active", "active")

reprodStages(matF, matFmu, post, maxRep, matrixStages)

# say what outputs you want
## combine postrep and rep
reprodStages(matF, matFmu, post, maxRep, matrixStages, includeProp = FALSE,
  includePost = FALSE)
## warn about prop?
reprodStages(matF, matFmu, post, maxRep, matrixStages, includeProp = FALSE)
## NOT ALLOWED
#reprodStages(matF, matFmu, post, maxRep, matrixStages, c('prop', 'rep', 'postrep'))
## NOT ALLOWED
#reprodStages(matF, matFmu, post, maxRep, matrixStages, c('prop', 'postrep'))
## NOT ALLOWED
#reprodStages(matF, matFmu, post, maxRep, matrixStages, c('prerep', 'postrep'))
## NOT ALLOWED
#reprodStages(matF, matFmu, post, maxRep, matrixStages, c('prop', 'prerep'))
}
